pyproject.toml
[project]
name = "pydantic-settings-azure-app-configuration"
version = "1.0.1"
requires-python = ">=3.12"
readme = "README.md"
description = "Azure App Configuration setting source for pydantic-settings"
dependencies = [
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.4",
    "azure-appconfiguration>=1.6.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-identity>=1.17.1",
]

[project.urls]
Repository = "https://github.com/AndreuCodina/pydantic-settings-azure-app-configuration"
Documentation = "https://github.com/AndreuCodina/pydantic-settings-azure-app-configuration"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_settings_azure_app_configuration"]

[tool.pyright]
typeCheckingMode = "strict"

[tool.ruff]
src = ["src", "tests"]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["F", "E", "W", "N", "I", "ANN"]
ignore = ["E501", "ANN101", "ANN102", "ANN401"]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
